refactor: move theme into separate javascript file
