+++ src/App.js
@@ -6,7 +6,7 @@ import 'leaflet.gridlayer.googlemutant';
 import Fade from "@material-ui/core/Fade";
 import Grid from '@material-ui/core/Grid';
 import Snackbar from "@material-ui/core/Snackbar";
-import { createMuiTheme,makeStyles, ThemeProvider } from "@material-ui/core/styles";
+import { makeStyles, ThemeProvider } from "@material-ui/core/styles";
 import MuiAlert from "@material-ui/lab/Alert";
 import expect from "expect-runtime";
 import log from "loglevel";
@@ -21,34 +21,9 @@ import getLogoSrc from "./models/logo";
 import Map from "./models/Map";
 import {parseMapName} from "./utils";
 
+import appTheme from "./theme";
 
 const MOBILE_WIDTH = 960;
-
-const PRIMARY = "#8bc34a"
-const SECONDARY = "#ffca28"
-
-const theme = createMuiTheme({
-	spacing		: 4,
-	typography		: {
-		fontFamily: [
-			'Roboto',
-			'Lato',
-			'Helvetica',
-			'Arial',
-			'sans-serif',
-		].join(','),
-	},
-  palette: {
-    primary: {
-      main: PRIMARY,
-    },
-    secondary: {
-      main: SECONDARY,
-    }
-  },
-});
-
-log.info("theme:", theme);
 
 const useStyles = makeStyles(theme => ({
   mapContainer: {
@@ -454,7 +429,7 @@ function App() {
   },[]);
 
   return (
-    <ThemeProvider theme={theme}>
+    <ThemeProvider theme={appTheme}>
       <SidePanel
         tree={tree}
         state={sidePanelState}
@@ -519,9 +494,4 @@ function App() {
   );
 }
 
-export {
-  PRIMARY,
-  SECONDARY,
-  theme,
-}
 export default App;
